requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asyncio==3.4.3
attrs==25.3.0
autopep8==2.3.2
certifi==2025.1.31
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dotenv-python==0.0.1
fastapi==0.105.0
flake8==7.2.0
fonttools==4.57.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
multidict==6.3.2
numpy==2.2.4
openai==1.71.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pycodestyle==2.13.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyflakes==3.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.24.0
yapf==0.43.0
yarl==1.19.0